10b/pyproject.toml
[tool.poetry]
name = "10b"
version = "0.1.0"
description = ""
authors = ["Abdi Abdulle <cph-aa394@cphbusiness.dk>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
stripe = "^9.7.0"
fastapi = "^0.111.0"
uvicorn = "^0.29.0"
python-dotenv = "^1.0.1"
aiofiles = "^23.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
